% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustInd_hierarch}
\alias{clustInd_hierarch}
\title{Perform hierarchical clustering for a different combinations of indexes,
method and distance}
\usage{
clustInd_hierarch(
  ind_data,
  vars_list,
  name_vars = NULL,
  method_list = c("single", "complete", "average", "centroid", "ward.D2"),
  dist_list = c("euclidean", "manhattan"),
  n_cluster = 2,
  true_labels = NULL,
  colapse = FALSE,
  num_cores = 1
)
}
\arguments{
\item{ind_data}{Dataframe containing indexes applied to the main data and
derivatives}

\item{vars_list}{List of variable sets for
clustering}

\item{name_vars}{A vector with names for \code{vars_list}. NULL by default
in which case names are set to vars1, ..., varsk, where k is the number of
elements in \code{vars_list}.}

\item{method_list}{List of clustering methods}

\item{dist_list}{List of distance metrics}

\item{n_cluster}{Number of clusters to create}

\item{true_labels}{Vector of true labels for validation
(if it is not known true_labels is set to NULL)}

\item{colapse}{It is a boolean. If it is true a dataframe with metrics values
is generated. If \code{true_labels} is True the dataframe contains Purity,
F-measure, RI and Time, and if it is False, only Time.}

\item{num_cores}{Number of cores to do parallel computation. 1 by default,
which mean no parallel execution.}
}
\value{
A list containing hierarchical clustering results
for each configuration
}
\description{
Perform hierarchical clustering for a different combinations of indexes,
method and distance
}
\examples{
vars1 <- c("dtaEI", "dtaMEI")
vars2 <- c("dtaHI", "dtaMHI")
data <- ehymet::sim_model_ex1()
data_ind <- ehymet::ind(data, t=seq(0, 1, length = 30))
clustInd_hierarch(data_ind, list(vars1, vars2))
}
