% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustInd_spc}
\alias{clustInd_spc}
\title{Perform spectral clustering for a different combinations of indexes
and kernels}
\usage{
clustInd_spc(
  ind_data,
  vars_list,
  kernel_list = c("rbfdot", "polydot"),
  n_cluster = 2,
  true_labels = NULL,
  ...
)
}
\arguments{
\item{ind_data}{Dataframe containing indexes applied to the main data
and derivatives}

\item{vars_list}{List of characters representing variable sets for clustering}

\item{kernel_list}{List of kernels}

\item{n_cluster}{Number of clusters to create}

\item{true_labels}{Vector of true labels for validation
(if it is not known true_labels is set to NULL)}

\item{...}{Additional arguments (unused)}
}
\value{
A list containing kkmeans clustering results for each configuration
}
\description{
Perform spectral clustering for a different combinations of indexes
and kernels
}
