% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustInd_svc}
\alias{clustInd_svc}
\title{Perform support vector clustering for a different combinations of indexes and distance}
\usage{
clustInd_svc(
  ind_data,
  vars_list,
  method_list = c("kmeans", "kernkmeans"),
  n_cluster = 2,
  true_labels = NULL,
  ...
)
}
\arguments{
\item{ind_data}{Dataframe containing indexes applied to the main data and derivatives}

\item{vars_list}{List of characters representing variable sets for clustering}

\item{method_list}{List of methods}

\item{n_cluster}{Number of clusters to create}

\item{true_labels}{Vector of true labels for validation (if it is not known true_labels is set to NULL)}

\item{...}{Additional arguments (unused)}
}
\value{
A list containing kkmeans clustering results for each configuration
}
\description{
Perform support vector clustering for a different combinations of indexes and distance
}
