% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_def.R
\name{EI}
\alias{EI}
\title{Epigraph Index (EI) for a univariate functional dataset.}
\usage{
EI(curves)
}
\arguments{
\item{curves}{A matrix where each row represents a curve, and each column
represents values along the curve.}
}
\value{
Return a numeric vector containing the EI for each curve
}
\description{
The Epigraph Index of a curve x is one minus the proportion of curves in the sample
that are above x.
}
\examples{
x <- matrix(c(1,2,3,3,2,1,5,2,3,9,8,7),ncol = 3, nrow = 4)
EI(x)
}
