% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ind}
\alias{ind}
\title{Create a dataset with indexes from a functional dataset in one or multiple
dimensions}
\usage{
ind(
  curves,
  grid_ll = 0,
  grid_ul = 1,
  nbasis = 25,
  norder = 4,
  indices = c("EI", "HI", "MEI", "MHI"),
  ...
)
}
\arguments{
\item{curves}{\code{matrix} with dimension \eqn{n \times p} in the case of a
one-dimensional functional dataset, or \code{array} of dimension
\eqn{n \times p \times k} in the case of a multivariate functional dataset.
\eqn{n} represents the number of curves, \eqn{p} the number of values along
the curve, and in the second case, \eqn{k} is the number of dimensions.}

\item{grid_ll}{lower limit of the grid.}

\item{grid_ul}{upper limit of the grid.}

\item{nbasis}{Number of basis for the B-splines}

\item{norder}{Order of the B-splines}

\item{indices}{Set of indices to be applied to the dataset. They should be
any between EI, HI, MEI and MHI}

\item{...}{Additional arguments (unused)}
}
\value{
A dataframe containing the indexes provided in \code{indices} for
original data, first and second derivatives
}
\description{
Create a dataset with indexes from a functional dataset in one or multiple
dimensions
}
\examples{
x1 <- array(c(1,2,3, 3,2,1, 5,2,3, 9,8,7, -1,-5,-6, 2,3,0, -1,0,2, -1,-2,0),
dim = c(3,4,2))
ind(x1, nbasis = 4)

x2 <- matrix(c(1,2,3,3,2,1,5,2,3,9,8,7), nrow = 3, ncol  = 4)
ind(x2, nbasis = 4)
}
