% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{mulMHI}
\alias{mulMHI}
\title{Modified hypograph Index (MHI) for a multivariate functional dataset.}
\usage{
mulMHI(curves)
}
\arguments{
\item{curves}{An array with dimension \eqn{n \times p \times k} with
\eqn{n} curves, \eqn{p} values along the curve, and \eqn{k} dimensions.}
}
\value{
A numeric vector containing the MHI for each multivariate curve
}
\description{
Modified hypograph Index (MHI) for a multivariate functional dataset.
}
\examples{
x <- array(c(1,2,3, 3,2,1, 5,2,3, 9,8,7, -1,-5,-6, 2,3,0, -1,0,2, -1,-2,0),
dim = c(3,4,2))
mulMHI(x)
}
