% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_models.R
\name{plt_fun}
\alias{plt_fun}
\title{Function for generating plots of one dimensional functional datasets}
\usage{
plt_fun(data, true_labels)
}
\arguments{
\item{data}{data matrix of size \eqn{n \times p}}

\item{true_labels}{array containing the true groups in which the data should
be classified}
}
\value{
A plot
}
\description{
Function for generating plots of one dimensional functional datasets
}
\examples{
dat1 <- sim_model_ex1()
true_labels <- c(rep(1,50), rep(2,50))
plt_fun(dat1, true_labels)
}
