% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustInd_hierarch}
\alias{clustInd_hierarch}
\title{Perform hierarchical clustering for a different combinations of indexes,
method and distance}
\usage{
clustInd_hierarch(
  ind_data,
  vars_list,
  method_list = c("single", "complete", "average", "centroid", "ward.D2"),
  dist_list = c("euclidean", "manhattan"),
  n_cluster = 2,
  true_labels = NULL
)
}
\arguments{
\item{ind_data}{Dataframe containing indexes applied to the main data and
derivatives}

\item{vars_list}{List of characters representing variable sets for
clustering}

\item{method_list}{List of clustering methods}

\item{dist_list}{List of distance metrics}

\item{n_cluster}{Number of clusters to create}

\item{true_labels}{Vector of true labels for validation
(if it is not known true_labels is set to NULL)}
}
\value{
A list containing hierarchical clustering results
for each configuration
}
\description{
Perform hierarchical clustering for a different combinations of indexes,
method and distance
}
