% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustInd_spc}
\alias{clustInd_spc}
\title{Spectral clustering using indices}
\usage{
clustInd_spc(
  ind_data,
  vars_combinations,
  kernel_list = c("rbfdot", "polydot"),
  n_cluster = 2,
  true_labels = NULL,
  n_cores = 1
)
}
\arguments{
\item{ind_data}{Dataframe containing indexes applied to the original data and
its first and second derivatives. See \link{generate_indices}.}

\item{vars_combinations}{\code{list} containing one or more combinations of indexes in
\code{ind_data}. If it is non-named, the names of the variables are set to
vars1, ..., varsk, where k is the number of elements in \code{vars_combinations}.}

\item{kernel_list}{List of kernels}

\item{n_cluster}{Number of clusters to create}

\item{true_labels}{Vector of true labels for validation
(if it is not known true_labels is set to NULL)}

\item{n_cores}{Number of cores to do parallel computation. 1 by default,
which mean no parallel execution.}
}
\value{
A list containing kkmeans clustering results for each configuration
}
\description{
Perform spectral clustering for a different combinations of indexes
and kernels
}
\examples{
vars1 <- c("dtaEI", "dtaMEI")
vars2 <- c("dtaHI", "dtaMHI")
data <- ehymet::sim_model_ex1()
data_ind <- generate_indices(data)
clustInd_spc(data_ind, list(vars1, vars2))

}
