% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHyClus.R
\name{EHyClus}
\alias{EHyClus}
\title{EHyClus method for clustering. It creates a multivariate dataset containing
the epigraph, hypograph and/or modified version on the curves and derivatives
and perform hierarchical clustering, kmeans, kernel kmeans, support vector
clustering and spectral clustering}
\usage{
EHyClus(
  curves,
  vars_combinations,
  nbasis = 30,
  n_clusters = 2,
  norder = 4,
  clustering_methods = c("hierarch", "kmeans", "kkmeans", "spc"),
  indices = c("EI", "HI", "MEI", "MHI"),
  l_method_hierarch = c("single", "complete", "average", "centroid", "ward.D2"),
  l_dist_hierarch = c("euclidean", "manhattan"),
  l_dist_kmeans = c("euclidean", "mahalanobis"),
  l_kernel = c("rbfdot", "polydot"),
  grid_ll = 0,
  grid_ul = 1,
  true_labels = NULL,
  verbose = FALSE,
  num_cores = 1,
  ...
)
}
\arguments{
\item{curves}{Dataset containing the curves to apply a clustering algorithm.
The functional dataset can be one dimensional (\eqn{n \times p}) where n is the number of
curves and p the number of time points, or multidimensional (\eqn{n \times p \times k}) where k
represents the number of dimensions in the data}

\item{vars_combinations}{\code{list} containing one or more combinations of variables.
Each element of the list should be an atomic \code{vector} of strings with the
names of the variables. Combinations with non-valid variable names will be discarded.
If it is non-named, the names of the variables are set to
vars1, ..., varsk, where k is the number of elements in \code{vars_combinations}.}

\item{nbasis}{Number of basis for the B-splines.}

\item{n_clusters}{Number of clusters to create}

\item{norder}{Order of the B-splines.}

\item{clustering_methods}{character vector specifying at least one of the following
clustering methods to be computed: "hierarch", "kmeans", "kkmeans", "spc".}

\item{indices}{Names of the indices that need to be generated. They should be
one or more between 'EI', 'HI', 'MEI' and 'MHI'. Depending on the dimension on the data
they are calculated for one or multiple dimension}

\item{l_method_hierarch}{\code{list} of clustering methods for hierarchical
clustering}

\item{l_dist_hierarch}{\code{list} of distances for hierarchical clustering}

\item{l_dist_kmeans}{\code{list} of distances for kmeans clustering}

\item{l_kernel}{\code{list} of kernels}

\item{grid_ll}{lower limit of the grid.}

\item{grid_ul}{upper limit of the grid.}

\item{true_labels}{Vector of true labels for validation
(if it is not known true_labels is set to NULL)}

\item{verbose}{If \code{TRUE}, the function will print logs for about the execution of
some clustering methods. Defaults to \code{FALSE}.}

\item{num_cores}{Number of cores to do parallel computation. 1 by default,
which mean no parallel execution.}

\item{...}{Ignored.}
}
\value{
A list containing the clustering partition for each method and indexes
combination and a data frame containing the time elapsed for obtaining a
clustering partition of the indexes dataset for each methodology
}
\description{
EHyClus method for clustering. It creates a multivariate dataset containing
the epigraph, hypograph and/or modified version on the curves and derivatives
and perform hierarchical clustering, kmeans, kernel kmeans, support vector
clustering and spectral clustering
}
\examples{
vars1 <- c("dtaEI", "dtaMEI"); vars2 <- c("dtaHI", "dtaMHI")
varsl <- list(vars1, vars2)
data <- sim_model_ex1()
EHyClus(data, varsl, grid_ll = 0, grid_ul = 1)

}
