% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_models.R
\name{sim_model_ex2}
\alias{sim_model_ex2}
\title{Function for generating functional data in one or multiple dimension}
\usage{
sim_model_ex2(n = 50, p = 150, i_sim = 1, seed = NULL)
}
\arguments{
\item{n}{Number of curves to generate for each of the two groups. Set to 50 by default.}

\item{p}{Number of grid points of the curves.
Curves are generated over the interval \eqn{[0, 1]}. Set to 150 grid point by default.}

\item{i_sim}{Integer set to \eqn{1, \ldots, 4}}

\item{seed}{A seed to set for reproducibility.
NULL by default in which case a seed is not set.}
}
\value{
data matrix of size \eqn{2n \times p} if \eqn{\code{i_sim} \in {1,2}}
or an array of dimensions
\eqn{2n \times p \times 2} if \eqn{\code{i_sim} \in {3, 4}}
}
\description{
Function for generating functional data in one or multiple dimension
}
\examples{
sm1 <- sim_model_ex2()
dim(sm1)

sm4 <- sim_model_ex2(i_sim=4)
dim(sm4)

}
